do not reload index.jsp after every request to the server
init jquery datatable on browser load (rm of onload attr from body)

use min versions of jquery and jquery plugins

add images dir necesseray for datatable plugin

refactoring of server side classes and return codes and content types

refactoring of client side js

rm of jstl tag library (from pom.xml as well)

+++ pressure-blood-web/src/main/java/pb/db/PressureBloodDAO.java
@@ -55,36 +55,30 @@ public class PressureBloodDAO {
 		}
 	}
 
-	public long deleteMeasure(String measureId, String remoteUser) {
+	public boolean deleteMeasure(String measureId, String remoteUser) {
 		MeasurementValidator.validateMeasureId(measureId);
 
 		EntityManager em = emf.createEntityManager();
 		Measurement measure = em.find(Measurement.class,
 				Long.parseLong(measureId));
 		try {
-			// if the entity does not exist
-			if (measure == null) {
-				return 0L;
-			} else {
-				// if the entity exists but does not belong to the current user
-				String measureUsername = measure.getUser().getUsername();
-				if (measureUsername.equals(remoteUser)) {
-					EntityTransaction et = em.getTransaction();
-					try {
-						et.begin();
-						em.remove(measure);
-						et.commit();
-					} finally {
-						if (et.isActive()) {
-							et.rollback();
-						}
+			if (measure != null
+					&& measure.getUser().getUsername().equals(remoteUser)) {
+				EntityTransaction et = em.getTransaction();
+				try {
+					et.begin();
+					em.remove(measure);
+					et.commit();
+				} finally {
+					if (et.isActive()) {
+						et.rollback();
 					}
-					return measure.getId();
 				}
+				return true;
 			}
 		} finally {
 			em.close();
 		}
-		return 0L;
+		return false;
 	}
 }
